Reorganized folders & renamed assets, "hardware controller" replaced with "serial input", moved logic out of `InputService`
