+++ SashimiSlayer/Assets/Data/Events/Input/Serial.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c3f8c3e0a36cd0647be50ffdcbc0a104
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
