+++ SashimiSlayer/Assets/Data/Events/Input/Serial/OnSerialConnect.asset
@@ -9,7 +9,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a2f70f35c4824ee09fcb50fd8385c0c7, type: 3}
-  m_Name: SetUseHardwareController
+  m_Script: {fileID: 11500000, guid: 509f402d747b42059ba177f6fdb1f99e, type: 3}
+  m_Name: OnSerialConnect
   m_EditorClassIdentifier: 
-  _description: Set whether to use the custom alt-control sword or not
+  _description: 'Triggers on connecting to serial input '
+  _debug: 0
