+++ SashimiSlayer/Assets/Data/Events/Input/Serial/OnSerialConnect.asset.meta
