+++ SashimiSlayer/Assets/Data/Events/Input/Serial/SerialConnectionStatus.asset.meta
