+++ SashimiSlayer/Assets/Data/Events/Input/Serial/UpAxisChanged.asset.meta
