+++ SashimiSlayer/Assets/Scripts/GameInput/SwordSerialReader.cs.meta
