+++ SashimiSlayer/Assets/Scripts/Menus/PauseMenu/Views/ControlDeviceSettingsView.cs
@@ -53,7 +53,7 @@ namespace Menus.PauseMenu.Views
         {
             _lastPortName = PlayerPrefs.GetString(LastPortName, "");
             _inputDeviceDropdown.ClearOptions();
-            _inputDeviceDropdown.AddOptions(new List<string> { "Conventional Controls", "Alt-Control Sword" });
+            _inputDeviceDropdown.AddOptions(new List<string> { "HID Input (Default)", "Sword Control Serial" });
 
             _inputDeviceDropdown.onValueChanged.AddListener(HandleInputDeviceChanged);
             _connectButton.onClick.AddListener(HandleConnect);
