+++ SashimiSlayer/Assets/Scripts/Menus/StartMenu/SliceInputHook.cs
@@ -1,4 +1,5 @@
 using EditorUtils.BoldHeader;
+using GameInput;
 using NaughtyAttributes;
 using UnityEngine;
 using UnityEngine.Events;
