+++ SashimiSlayer/Assets/Scripts/Menus/StartMenu/StartMenuButton.cs
@@ -1,5 +1,6 @@
 using DG.Tweening;
 using EditorUtils.BoldHeader;
+using GameInput;
 using NaughtyAttributes;
 using UnityEngine;
 using UnityEngine.Events;
